add defautl salary cap showdown generator

+++ DraftKings.LineupGenerator.Models/Draftables/DraftableDisplayModel.cs
@@ -6,11 +6,11 @@
 
         public decimal Fppg { get; set; }
 
-        public int Salary { get; set; }
+        public decimal Salary { get; set; }
 
         public string RosterPosition { get; set; }
 
-        public DraftableDisplayModel(string name, decimal fppg, int salary, string rosterPosition)
+        public DraftableDisplayModel(string name, decimal fppg, decimal salary, string rosterPosition)
         {
             Name = name;
             Fppg = fppg;
